Formatting

+++ src/features/cart/cartSlice.jsx
@@ -5,8 +5,8 @@ const initialState = {
 };
 
 const cartSlice = createSlice({
-  initialState,
   name: "cart",
+  initialState,
   reducers: {
     addItem(state, action) {
       state.cart.push(action.payload);
@@ -23,9 +23,7 @@ const cartSlice = createSlice({
       const item = state.cart.find((item) => item.pizzaId === action.payload);
       item.quantity--;
       item.totalPrice = item.quantity * item.unitPrice;
-      if (item.quantity === 0) {
-        cartSlice.caseReducers.deleteItem(state, action);
-      }
+      if (item.quantity === 0) cartSlice.caseReducers.deleteItem(state, action);
     },
     clearCart(state) {
       state.cart = [];
@@ -45,14 +43,11 @@ export default cartSlice.reducer;
 
 export const getCart = (state) => state.cart.cart;
 
-export const getNumItems = (state) => {
-  return state.cart.cart.reduce((sum, item) => sum + item.quantity, 0);
-};
+export const getTotalCartQuantity = (state) =>
+  state.cart.cart.reduce((sum, item) => sum + item.quantity, 0);
 
-export const getTotalPrice = (state) => {
-  return state.cart.cart.reduce((sum, item) => sum + item.totalPrice, 0);
-};
+export const getTotalCartPrice = (state) =>
+  state.cart.cart.reduce((sum, item) => sum + item.totalPrice, 0);
 
-export const getCurrentQuantity = (id) => (state) => {
-  return state.cart.cart.find((item) => item.pizzaId === id)?.quantity ?? 0;
-};
+export const getCurrentQuantityById = (id) => (state) =>
+  state.cart.cart.find((item) => item.pizzaId === id)?.quantity ?? 0;
